feat: support namespace based claims (#485)
When adding a custom claim in Auth0 via "Actions", you must add a namespace. According to the standard this needs to be a https URI. The existing logic breaks due to the path splitting logic.
This enhancement fixes that by first checking the claim like configured in the options with https://... and then falling back to the original logic.